Move cleanup to namespace_ctl and cleanup on PAUSE
